Local models are now created on-the-fly
Instead of creating the models when initializing the federated learning
clients, the models are now created on-the-fly when the local training
is started. Although this is a little slower, it is much more memory
efficient, because there will only ever be three models in RAM/VRAM: the
global model, the local model of the current client, and the aggregation
of the the client models. This way, an arbitrary number of clients can
be trained used.