Bug fix in the plot-training-statistics command
The final validation loss and accuracy of the federated learning
experiments were displayed in exponential notation, but they should
have been displayed as a regular float with 2 decimal digits. Also added
some more logging to let the user know what is going on, because some
parts of the plotting were pretty slow.